pyproject.toml
[project]
name = "InsightFlow-AI"
version = "0.1.0"
description = "InsightFlow AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv==2.1.3",
    "beautifulsoup4==4.13.3",
    "chainlit==2.5.5",
    "cohere==5.13.12",
    "datasets==3.3.1",
    "faiss-cpu==1.10.0",
    "langchain-cohere==0.4.2",
    "langchain-community==0.3.14",
    "langchain-huggingface==0.1.2",
    "langchain-openai==0.2.14",
    "langchain-qdrant==0.2.0",
    "langgraph==0.2.61",
    "lxml==5.3.1",
    "nltk==3.8.1",
    "numpy>=1.25.0,<2.0.0",
    "pyarrow==19.0.1",
    "pymupdf==1.25.3",
    "python-dotenv>=1.0.1",
    "python-pptx==1.0.2",
    "ragas==0.2.10",
    "sentence-transformers==3.4.1",
    "unstructured==0.14.8",
    "websockets>=15.0",
    "fpdf==1.7.2",
    "qdrant-client",
    "tiktoken",
    "fastembed",
]

[tool.pytest.ini_options]
python_files = "test_*.py tests_*.py *_test.py *_tests.py"
testpaths = [
    "tests",
]
pythonpath = ["."]
asyncio_mode = "auto"
